Fixed unexpected lists on HTTP post inputs
When handling standard HTTP post requests, the POST data is converted to a dictionary in order to standardize the input between Websocket and HTTP methods. In converting the postdata to a dictionary, it put in every value corresponding to a key in the post data as a list, even if it only has one input value.
The only elements that should be a list, are elements that could expect more than 1 value. All other elements should only expect 1 single value.